third_party/blink/web_tests/platform/generic/fast/dom/htmlcollection-length-after-item-expected.txt

This tests accessing the length after accessing (length + 1)-th item in HTMLCollection doesn't cache a wrong length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS children = container.children; children.length is 0
PASS container.appendChild(span); children[1]; children.length is 1
PASS container.removeChild(span); children.length is 0
PASS successfullyParsed is true

TEST COMPLETE
